Extract user management to User class
