+++ src/AJAX.php
@@ -75,111 +75,24 @@ class AJAX {
 		}
 
 		$id_token = $this->jwt->get_payload( $tokens['id_token'] );
-
-		// The following fields are optional, and must have a default value.
-		$id_token = wp_parse_args(
-			$id_token,
-			array(
-				'locale' => str_replace( '-', '_', get_locale() ),
-			)
-		);
-
-		$userdata = array(
-			'role'        => 'customer',
-			'user_login'  => sanitize_user( $id_token['email'] ),
-			'user_pass'   => wp_generate_password(),
-			'user_email'  => sanitize_email( $id_token['email'] ),
-			'first_name'  => sanitize_text_field( $id_token['given_name'] ),
-			'last_name'   => sanitize_text_field( $id_token['family_name'] ),
-			'description' => __( 'Sign in with Klarna', 'siwk' ),
-			'locale'      => $id_token['locale'],
-		);
-
-		// Clean fields, and use default values to avoid undefined index.
-		$billing_address = array_map(
-			function ( $field ) {
-				if ( empty( $field ) ) {
-					return '';
-				}
-				return wc_clean( $field );
-			},
-			$id_token['billing_address']
-		);
-
-		$userdata['meta_input'] = array(
-			'billing_first_name'    => $userdata['first_name'],
-			'billing_last_name'     => $userdata['last_name'],
-			'billing_city'          => $billing_address['city'],
-			'billing_state'         => $billing_address['region'],
-			'billing_country'       => $billing_address['country'],
-			'billing_postcode'      => $billing_address['postal_code'],
-			'billing_address_1'     => $billing_address['street_address'],
-			'billing_address_2'     => $billing_address['street_address_2'],
-			'billing_phone'         => $id_token['phone'],
-			'billing_email'         => $userdata['user_email'],
-			'shipping_first_name'   => $userdata['first_name'],
-			'shipping_last_name'    => $userdata['last_name'],
-			'shipping_city'         => $billing_address['city'],
-			'shipping_country'      => $billing_address['country'],
-			'shipping_state'        => $billing_address['region'],
-			'shipping_postcode'     => $billing_address['postal_code'],
-			'shipping_address_1'    => $billing_address['street_address'],
-			'shipping_address_2'    => $billing_address['street_address_2'],
-			'shipping_phone'        => $id_token['phone'],
-			'shipping_email'        => $userdata['user_email'],
-			User::REFRESH_TOKEN_KEY => $refresh_token,
-		);
-
-		// Remove empty fields (based on default value).
-		$userdata['meta_input'] = array_filter(
-			$userdata['meta_input'],
-			function ( $field ) {
-				return ! empty( $field );
-			}
-		);
-
-		// If the user is already logged in, save the refresh token (to be used for klarna_access_token), and return.
-		// Otherwise, their cart will be emptied when changing account.
-		$user_id = get_current_user_id();
-		$guest   = 0;
-		if ( $guest !== $user_id ) {
-			$this->user->set_tokens( $user_id, $tokens );
-			$this->user->set_refresh_token( $user_id, $refresh_token );
-
-			wp_send_json_success( 'user already logged in' );
-		}
+		$userdata = $this->user->get_user_data( $id_token, $refresh_token );
 
 		if ( username_exists( $userdata['user_login'] ) || email_exists( $userdata['user_email'] ) ) {
-			$user = get_user_by( 'login', $userdata['user_login'] );
-			$user = ! empty( $user ) ? $user : get_user_by( 'email', $userdata['user_email'] );
-			if ( empty( $user ) ) {
-				wp_send_json_error( 'user exists, failed to retrieve user data' );
-			}
-
-			// Skip if the user is already logged in.
-			if ( get_current_user_id() === $user->ID ) {
-				wp_send_json_success( 'user exists, already logged in' );
-			}
-
-			// Try to log the user in. The client should refresh the page.
-			$this->user->set_current_user( $user->ID );
-			$this->user->set_tokens( $user->ID, $tokens );
-
-			do_action( 'siwk_existing_user_logged_in', $user->ID, $userdata );
-			wp_send_json_success( 'user exists, logging in' );
+			$result = $this->user->merge_with_existing_user( $userdata );
+		} else {
+			$result = $this->user->register_new_user( $userdata );
 		}
 
-		$user_id = wp_insert_user( apply_filters( 'siwk_create_new_user', $userdata ) );
-		if ( is_wp_error( $user_id ) ) {
-			wp_send_json_error( 'could not create user' );
+		if ( is_wp_error( $result ) ) {
+			wp_send_json_error( $result->get_error_message() );
 		}
 
-		do_action( 'woocommerce_created_customer', $user_id, $userdata, false );
+		$guest   = 0;
+		$user_id = get_current_user_id();
+		if ( $guest !== $user_id ) {
+			$this->user->sign_in_user( $user_id, $tokens, $refresh_token );
+		}
 
-		// Try to log the user in. The page should be automatically refreshed in the client.
-		$this->user->set_current_user( $user_id );
-		$this->user->set_tokens( $user_id, $tokens );
-
-		wp_send_json_success( 'user created, logging in' );
+		wp_send_json_success( array( 'user_id' => $user_id ) );
 	}
 }
